아이디 받기 전 X 누를 때 오류 수정

+++ src/BetweenUser.js
@@ -18,8 +18,10 @@ const BetweenUser = () => {
 
   const reload = async () => {
     try {
-      const response = await axios.delete(`https://yunharyu.shop/api/interactions/${myID}`)
-    
+      if(myID) {
+        const response = await axios.delete(`https://yunharyu.shop/api/interactions/${myID}`)
+      }
+      
       nav("/");
       
     } catch (error) {
